News Bot/requirements.txt
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
decorator==4.4.2
dialogflow==1.1.0
feedparser==6.0.1
Flask==1.1.2
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.22.2
google-auth==1.21.2
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
rsa==4.6
sgmllib3k==1.0.0
six==1.15.0
tornado==6.0.4
urllib3==1.25.10
Werkzeug==1.0.1
